+++ spec/chatgpt_cli_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-RSpec.describe OpenAIPipe do
+RSpec.describe ChatGPTCLI do
   it "has a version number" do
-    expect(OpenAIPipe::VERSION).not_to be nil
+    expect(ChatGPTCLI::VERSION).not_to be nil
   end
 
   it "does something useful" do