+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## [0.1.0] - 2022-12-23
 
 - Initial release
+
+## [0.1.1] - 2023-03-23
+
+- Define open ai model in env var
+- Input context pompts